+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New danari character customizations
 - Bald hairstyles for humans and danari
 - AI for sceptre weilders and sceptre cultists in Tier 5 dungeons
+- HUD debug info now displays current biome and site
 
 ### Changed